resolve some differences between parsed dicts and serialized dicts

+++ src/FoSpy/parsing/validators/reaction.py
@@ -1,6 +1,8 @@
+from decimal import Decimal
+
 def nominal_mass(val):
     try:
-        nominal_mass = float(val)
+        nominal_mass = Decimal(val)
     except:
         raise ValueError(f"Unable to convert 'nominal_mass: {val}' into a decimal number.")
     if nominal_mass > 0:
